# src/EdFi.RtI.Api/EdFi.RtI.Api.Swagger.yaml
openapi: 3.0.1
info:
  title: Response to Intervention API
  description: The Management HTTP API.
  termsOfService: https://edgraph.com/legal/termsofservices
  contact:
    name: EdGraph Support Team
    email: support@edgraph.com
  license:
    name: Use under EdGraph license
    url: https://edgraph.com/legal/license
  version: v1
servers:
  - url: http://docker.for.win.localhost:5001/rti
paths:
...
  /ScoringInterventions:
    post:
      tags:
        - ScoringInterventions
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/EdFi.RtI.Core.RequestBodies.StudentInterventionBody'
          application/json:
            schema:
              $ref: '#/components/schemas/EdFi.RtI.Core.RequestBodies.StudentInterventionBody'
          text/json:
            schema:
              $ref: '#/components/schemas/EdFi.RtI.Core.RequestBodies.StudentInterventionBody'
          application/*+json:
            schema:
              $ref: '#/components/schemas/EdFi.RtI.Core.RequestBodies.StudentInterventionBody'
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.StudentInterventionAssociation'
            application/json:
              schema:
                $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.StudentInterventionAssociation'
            text/json:
              schema:
                $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.StudentInterventionAssociation'
  /ScoringInterventions/delete:
    put:
      tags:
        - ScoringInterventions
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.StudentInterventionAssociation'
          application/json:
            schema:
              $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.StudentInterventionAssociation'
          text/json:
            schema:
              $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.StudentInterventionAssociation'
          application/*+json:
            schema:
              $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.StudentInterventionAssociation'
      responses:
        '200':
          description: Success
  /ScoringInterventions/scorings:
    post:
      tags:
        - ScoringInterventions
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/EdFi.RtI.Core.RequestBodies.InterventionScoringFiltersBody'
          application/json:
            schema:
              $ref: '#/components/schemas/EdFi.RtI.Core.RequestBodies.InterventionScoringFiltersBody'
          text/json:
            schema:
              $ref: '#/components/schemas/EdFi.RtI.Core.RequestBodies.InterventionScoringFiltersBody'
          application/*+json:
            schema:
              $ref: '#/components/schemas/EdFi.RtI.Core.RequestBodies.InterventionScoringFiltersBody'
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EdFi.RtI.Core.DTOs.Composed.StudentInterventionsDTO'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EdFi.RtI.Core.DTOs.Composed.StudentInterventionsDTO'
            text/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EdFi.RtI.Core.DTOs.Composed.StudentInterventionsDTO'
  '/ScoringInterventions/scorings/{studentUniqueId}':
    get:
      tags:
        - ScoringInterventions
      parameters:
        - name: studentUniqueId
          in: path
          required: true
          schema:
            type: string
            nullable: true
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/EdFi.RtI.Core.DTOs.Composed.StudentInterventionsDTO'
            application/json:
              schema:
                $ref: '#/components/schemas/EdFi.RtI.Core.DTOs.Composed.StudentInterventionsDTO'
            text/json:
              schema:
                $ref: '#/components/schemas/EdFi.RtI.Core.DTOs.Composed.StudentInterventionsDTO'
...
components:
  schemas:
...
    EdFi.RtI.Core.DTOs.Student.AddressesDTO:
      type: object
      additionalProperties: false
    EdFi.RtI.Core.DTOs.Student.CharacteristicsDTO:
      type: object
      additionalProperties: false
    EdFi.RtI.Core.DTOs.Student.CohortYearsDTO:
      type: object
      additionalProperties: false
    EdFi.RtI.Core.DTOs.Student.DisabilitiesDTO:
      type: object
      additionalProperties: false
...
    EdFi.RtI.Core.DTOs.Student.StudentDTO:
      type: object
      properties:
        _Etag:
          type: string
          nullable: true
        addresses:
          type: array
          items:
            $ref: '#/components/schemas/EdFi.RtI.Core.DTOs.Student.AddressesDTO'
          nullable: true
        birthCity:
          type: string
          nullable: true
        birthCountryDescriptor:
          type: string
          nullable: true
        birthDate:
          type: string
          format: date-time
        birthInternationalProvince:
          type: string
          nullable: true
        birthStateAbbreviationType:
          type: string
          nullable: true
        characteristics:
          type: array
          items:
            $ref: '#/components/schemas/EdFi.RtI.Core.DTOs.Student.CharacteristicsDTO'
          nullable: true
        citizenshipStatusType:
          type: string
          nullable: true
        cohortYears:
          type: array
          items:
            $ref: '#/components/schemas/EdFi.RtI.Core.DTOs.Student.CohortYearsDTO'
          nullable: true
        dateEnteredUS:
          type: string
          nullable: true
        disabilities:
          type: array
          items:
            $ref: '#/components/schemas/EdFi.RtI.Core.DTOs.Student.DisabilitiesDTO'
          nullable: true
        displacementStatus:
          type: string
          nullable: true
        economicDisadvantaged:
          type: string
          nullable: true
        electronicMails:
          type: array
          items:
            $ref: '#/components/schemas/EdFi.RtI.Core.DTOs.Student.ElectronicMailsDTO'
          nullable: true
        firstName:
          type: string
          nullable: true
        generationCodeSuffix:
          type: string
          nullable: true
        hispanicLatinoEthnicity:
          type: boolean
        id:
          type: string
          nullable: true
        identificationCodes:
          type: array
          items:
            $ref: '#/components/schemas/EdFi.RtI.Core.DTOs.Student.IdentificationCodesDTO'
          nullable: true
        identificationDocuments:
          type: array
          items:
            $ref: '#/components/schemas/EdFi.RtI.Core.DTOs.Student.IdentificationDocumentsDTO'
          nullable: true
        indicators:
          type: array
          items:
            $ref: '#/components/schemas/EdFi.RtI.Core.DTOs.Student.IndicatorsDTO'
          nullable: true
        internationalAddresses:
          type: array
          items:
            $ref: '#/components/schemas/EdFi.RtI.Core.DTOs.Student.InternationalAddressesDTO'
          nullable: true
        languages:
          type: array
          items:
            $ref: '#/components/schemas/EdFi.RtI.Core.DTOs.Student.LanguagesDTO'
          nullable: true
        lastSurname:
          type: string
          nullable: true
        learningStyle:
          type: string
          nullable: true
        limitedEnglishProficiencyDescriptor:
          type: string
          nullable: true
        loginId:
          type: string
          nullable: true
        maidenName:
          type: string
          nullable: true
        middleName:
          type: string
          nullable: true
        multipleBirthStatus:
          type: string
          nullable: true
        oldEthnicityType:
          type: string
          nullable: true
        otherNames:
          type: array
          items:
            $ref: '#/components/schemas/EdFi.RtI.Core.DTOs.Student.OtherNamesDTO'
          nullable: true
        personalTitlePrefix:
          type: string
          nullable: true
        profileThumbnail:
          type: string
          nullable: true
        programParticipations:
          type: array
          items:
            $ref: '#/components/schemas/EdFi.RtI.Core.DTOs.Student.ProgramParticipationsDTO'
          nullable: true
        races:
          type: array
          items:
            $ref: '#/components/schemas/EdFi.RtI.Core.DTOs.Student.RacesDTO'
          nullable: true
        schoolFoodServicesEligibilityDescriptor:
          type: string
          nullable: true
        sexType:
          type: string
          nullable: true
        studentUniqueId:
          type: string
          nullable: true
        telephones:
          type: array
          items:
            $ref: '#/components/schemas/EdFi.RtI.Core.DTOs.Student.TelephonesDTO'
          nullable: true
        visas:
          type: array
          items:
            $ref: '#/components/schemas/EdFi.RtI.Core.DTOs.Student.VisasDTO'
          nullable: true
        fullName:
          type: string
          nullable: true
      additionalProperties: false
...
    EdFi.RtI.Core.RequestBodies.InterventionScoringFiltersBody:
      type: object
      properties:
        staffUniqueId:
          type: string
          nullable: true
        sectionId:
          type: string
          nullable: true
        showInactiveStudents:
          type: boolean
      additionalProperties: false
...
    EdFi.Ods.Api.Client.Models.StudentAddressesItem:
      type: object
      properties:
        addressType:
          type: string
          nullable: true
        apartmentRoomSuiteNumber:
          type: string
          nullable: true
        beginDate:
          type: string
          format: date-time
          nullable: true
        buildingSiteNumber:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        countyFIPSCode:
          type: string
          nullable: true
        endDate:
          type: string
          format: date-time
          nullable: true
        latitude:
          type: string
          nullable: true
        longitude:
          type: string
          nullable: true
        nameOfCounty:
          type: string
          nullable: true
        postalCode:
          type: string
          nullable: true
        stateAbbreviationType:
          type: string
          nullable: true
        streetNumberName:
          type: string
          nullable: true
      additionalProperties: false
    EdFi.Ods.Api.Client.Models.StudentCharacteristicsItem:
      type: object
      properties:
        beginDate:
          type: string
          format: date-time
          nullable: true
        descriptor:
          type: string
          nullable: true
        designatedBy:
          type: string
          nullable: true
        endDate:
          type: string
          format: date-time
          nullable: true
      additionalProperties: false
...
    EdFi.Ods.Api.Client.Models.StudentCohortYearsItem:
      type: object
      properties:
        cohortYearType:
          type: string
          nullable: true
        schoolYearTypeReference:
          $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.StudentCohortYearsItemSchoolYearTypeReference'
      additionalProperties: false
    EdFi.Ods.Api.Client.Models.StudentDisabilitiesItem:
      type: object
      properties:
        disabilityDescriptor:
          type: string
          nullable: true
        disabilityDeterminationSourceType:
          type: string
          nullable: true
        disabilityDiagnosis:
          type: string
          nullable: true
        orderOfDisability:
          type: integer
          format: int32
          nullable: true
      additionalProperties: false
    EdFi.Ods.Api.Client.Models.StudentElectronicMailsItem:
      type: object
      properties:
        electronicMailAddress:
          type: string
          nullable: true
        electronicMailType:
          type: string
          nullable: true
        primaryEmailAddressIndicator:
          type: boolean
          nullable: true
      additionalProperties: false
    EdFi.Ods.Api.Client.Models.StudentIdentificationCodesItem:
      type: object
      properties:
        assigningOrganizationIdentificationCode:
          type: string
          nullable: true
        identificationCode:
          type: string
          nullable: true
        studentIdentificationSystemDescriptor:
          type: string
          nullable: true
      additionalProperties: false
    EdFi.Ods.Api.Client.Models.StudentIdentificationDocumentsItem:
      type: object
      properties:
        documentExpirationDate:
          type: string
          format: date-time
          nullable: true
        documentTitle:
          type: string
          nullable: true
        identificationDocumentUseType:
          type: string
          nullable: true
        issuerCountryDescriptor:
          type: string
          nullable: true
        issuerDocumentIdentificationCode:
          type: string
          nullable: true
        issuerName:
          type: string
          nullable: true
        personalInformationVerificationType:
          type: string
          nullable: true
      additionalProperties: false
    EdFi.Ods.Api.Client.Models.StudentIndicatorsItem:
      type: object
      properties:
        beginDate:
          type: string
          format: date-time
          nullable: true
        designatedBy:
          type: string
          nullable: true
        endDate:
          type: string
          format: date-time
          nullable: true
        indicator:
          type: string
          nullable: true
        indicatorGroup:
          type: string
          nullable: true
        indicatorName:
          type: string
          nullable: true
      additionalProperties: false
    EdFi.Ods.Api.Client.Models.StudentInternationalAddressesItem:
      type: object
      properties:
        addressLine1:
          type: string
          nullable: true
        addressLine2:
          type: string
          nullable: true
        addressLine3:
          type: string
          nullable: true
        addressLine4:
          type: string
          nullable: true
        addressType:
          type: string
          nullable: true
        beginDate:
          type: string
          format: date-time
          nullable: true
        countryDescriptor:
          type: string
          nullable: true
        endDate:
          type: string
          format: date-time
          nullable: true
        latitude:
          type: string
          nullable: true
        longitude:
          type: string
          nullable: true
      additionalProperties: false
...
    EdFi.Ods.Api.Client.Models.StudentLanguagesItem:
      type: object
      properties:
        languageDescriptor:
          type: string
          nullable: true
        uses:
          type: array
          items:
            $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.StudentLanguagesItemUsesItem'
          nullable: true
      additionalProperties: false
    EdFi.Ods.Api.Client.Models.StudentLearningStyle:
      type: object
      properties:
        auditoryLearning:
          type: number
          format: double
        tactileLearning:
          type: number
          format: double
        visualLearning:
          type: number
          format: double
      additionalProperties: false
    EdFi.Ods.Api.Client.Models.StudentOtherNamesItem:
      type: object
      properties:
        firstName:
          type: string
          nullable: true
        generationCodeSuffix:
          type: string
          nullable: true
        lastSurname:
          type: string
          nullable: true
        middleName:
          type: string
          nullable: true
        otherNameType:
          type: string
          nullable: true
        personalTitlePrefix:
          type: string
          nullable: true
      additionalProperties: false
...
    EdFi.Ods.Api.Client.Models.StudentProgramParticipationsItem:
      type: object
      properties:
        beginDate:
          type: string
          format: date-time
          nullable: true
        designatedBy:
          type: string
          nullable: true
        endDate:
          type: string
          format: date-time
          nullable: true
        programCharacteristics:
          type: array
          items:
            $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.StudentProgramParticipationsItemProgramCharacteristicsItem'
          nullable: true
        programType:
          type: string
          nullable: true
      additionalProperties: false
    EdFi.Ods.Api.Client.Models.StudentRacesItem:
      type: object
      properties:
        raceType:
          type: string
          nullable: true
      additionalProperties: false
    EdFi.Ods.Api.Client.Models.StudentTelephonesItem:
      type: object
      properties:
        orderOfPriority:
          type: integer
          format: int32
          nullable: true
        telephoneNumber:
          type: string
          nullable: true
        telephoneNumberType:
          type: string
          nullable: true
        textMessageCapabilityIndicator:
          type: boolean
          nullable: true
      additionalProperties: false
    EdFi.Ods.Api.Client.Models.StudentVisasItem:
      type: object
      properties:
        visaType:
          type: string
          nullable: true
      additionalProperties: false
    EdFi.Ods.Api.Client.Models.Student:
      type: object
      properties:
        _etag:
          type: string
          nullable: true
        addresses:
          type: array
          items:
            $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.StudentAddressesItem'
          nullable: true
        birthCity:
          type: string
          nullable: true
        birthCountryDescriptor:
          type: string
          nullable: true
        birthDate:
          type: string
          format: date-time
        birthInternationalProvince:
          type: string
          nullable: true
        birthStateAbbreviationType:
          type: string
          nullable: true
        characteristics:
          type: array
          items:
            $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.StudentCharacteristicsItem'
          nullable: true
        citizenshipStatusType:
          type: string
          nullable: true
        cohortYears:
          type: array
          items:
            $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.StudentCohortYearsItem'
          nullable: true
        dateEnteredUS:
          type: string
          format: date-time
          nullable: true
        disabilities:
          type: array
          items:
            $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.StudentDisabilitiesItem'
          nullable: true
        displacementStatus:
          type: string
          nullable: true
        economicDisadvantaged:
          type: boolean
          nullable: true
        electronicMails:
          type: array
          items:
            $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.StudentElectronicMailsItem'
          nullable: true
        firstName:
          type: string
          nullable: true
        generationCodeSuffix:
          type: string
          nullable: true
        hispanicLatinoEthnicity:
          type: boolean
        id:
          type: string
          nullable: true
        identificationCodes:
          type: array
          items:
            $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.StudentIdentificationCodesItem'
          nullable: true
        identificationDocuments:
          type: array
          items:
            $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.StudentIdentificationDocumentsItem'
          nullable: true
        indicators:
          type: array
          items:
            $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.StudentIndicatorsItem'
          nullable: true
        internationalAddresses:
          type: array
          items:
            $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.StudentInternationalAddressesItem'
          nullable: true
        languages:
          type: array
          items:
            $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.StudentLanguagesItem'
          nullable: true
        lastSurname:
          type: string
          nullable: true
        learningStyle:
          $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.StudentLearningStyle'
        limitedEnglishProficiencyDescriptor:
          type: string
          nullable: true
        loginId:
          type: string
          nullable: true
        maidenName:
          type: string
          nullable: true
        middleName:
          type: string
          nullable: true
        multipleBirthStatus:
          type: boolean
          nullable: true
        oldEthnicityType:
          type: string
          nullable: true
        otherNames:
          type: array
          items:
            $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.StudentOtherNamesItem'
          nullable: true
        personalTitlePrefix:
          type: string
          nullable: true
        profileThumbnail:
          type: string
          nullable: true
        programParticipations:
          type: array
          items:
            $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.StudentProgramParticipationsItem'
          nullable: true
        races:
          type: array
          items:
            $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.StudentRacesItem'
          nullable: true
        schoolFoodServicesEligibilityDescriptor:
          type: string
          nullable: true
        sexType:
          type: string
          nullable: true
        studentUniqueId:
          type: string
          nullable: true
        telephones:
          type: array
          items:
            $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.StudentTelephonesItem'
          nullable: true
        visas:
          type: array
          items:
            $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.StudentVisasItem'
          nullable: true
      additionalProperties: false
...
    EdFi.Ods.Api.Client.Models.InterventionAppropriateGradeLevelsItem:
      type: object
      properties:
        gradeLevelDescriptor:
          type: string
          nullable: true
      additionalProperties: false
    EdFi.Ods.Api.Client.Models.InterventionAppropriateSexesItem:
      type: object
      properties:
        sexType:
          type: string
          nullable: true
      additionalProperties: false
    EdFi.Ods.Api.Client.Models.InterventionDiagnosesItem:
      type: object
      properties:
        diagnosisDescriptor:
          type: string
          nullable: true
      additionalProperties: false
...
    EdFi.Ods.Api.Client.Models.InterventionEducationContentsItem:
      type: object
      properties:
        educationContentReference:
          $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.InterventionEducationContentsItemEducationContentReference'
      additionalProperties: false
...
    EdFi.Ods.Api.Client.Models.InterventionEducationOrganizationReference:
      type: object
      properties:
        educationOrganizationId:
          type: integer
          format: int32
        link:
          $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.InterventionEducationOrganizationReferenceLink'
      additionalProperties: false
...
    EdFi.Ods.Api.Client.Models.InterventionInterventionPrescriptionsItem:
      type: object
      properties:
        interventionPrescriptionReference:
          $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.InterventionInterventionPrescriptionsItemInterventionPrescriptionReference'
      additionalProperties: false
    EdFi.Ods.Api.Client.Models.InterventionLearningResourceMetadataURIsItem:
      type: object
      properties:
        learningResourceMetadataURI:
          type: string
          nullable: true
      additionalProperties: false
...
    EdFi.Ods.Api.Client.Models.InterventionMeetingTimesItem:
      type: object
      properties:
        alternateDayName:
          type: string
          nullable: true
        classPeriodReference:
          $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.InterventionMeetingTimesItemClassPeriodReference'
        endTime:
          type: string
          nullable: true
        officialAttendancePeriod:
          type: boolean
          nullable: true
        startTime:
          type: string
          nullable: true
      additionalProperties: false
    EdFi.Ods.Api.Client.Models.InterventionPopulationServedsItem:
      type: object
      properties:
        populationServedType:
          type: string
          nullable: true
      additionalProperties: false
...
    EdFi.Ods.Api.Client.Models.InterventionStaffsItem:
      type: object
      properties:
        staffReference:
          $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.InterventionStaffsItemStaffReference'
      additionalProperties: false
    EdFi.Ods.Api.Client.Models.InterventionUrisItem:
      type: object
      properties:
        uri:
          type: string
          nullable: true
      additionalProperties: false
    EdFi.Ods.Api.Client.Models.Intervention:
      type: object
      properties:
        _etag:
          type: string
          nullable: true
        appropriateGradeLevels:
          type: array
          items:
            $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.InterventionAppropriateGradeLevelsItem'
          nullable: true
        appropriateSexes:
          type: array
          items:
            $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.InterventionAppropriateSexesItem'
          nullable: true
        beginDate:
          type: string
          format: date-time
        classType:
          type: string
          nullable: true
        deliveryMethodType:
          type: string
          nullable: true
        diagnoses:
          type: array
          items:
            $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.InterventionDiagnosesItem'
          nullable: true
        educationContents:
          type: array
          items:
            $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.InterventionEducationContentsItem'
          nullable: true
        educationOrganizationReference:
          $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.InterventionEducationOrganizationReference'
        endDate:
          type: string
          format: date-time
          nullable: true
        id:
          type: string
          nullable: true
        identificationCode:
          type: string
          nullable: true
        interventionPrescriptions:
          type: array
          items:
            $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.InterventionInterventionPrescriptionsItem'
          nullable: true
        learningResourceMetadataURIs:
          type: array
          items:
            $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.InterventionLearningResourceMetadataURIsItem'
          nullable: true
        meetingTimes:
          type: array
          items:
            $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.InterventionMeetingTimesItem'
          nullable: true
        populationServeds:
          type: array
          items:
            $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.InterventionPopulationServedsItem'
          nullable: true
        staffs:
          type: array
          items:
            $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.InterventionStaffsItem'
          nullable: true
        uris:
          type: array
          items:
            $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.InterventionUrisItem'
          nullable: true
      additionalProperties: false
...
    EdFi.RtI.Core.RequestBodies.StudentInterventionBody:
      type: object
      properties:
        student:
          $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.Student'
        intervention:
          $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.Intervention'
      additionalProperties: false
    EdFi.Ods.Api.Client.Models.StudentInterventionAssociationCohortReferenceLink:
      type: object
      properties:
        href:
          type: string
          nullable: true
        rel:
          type: string
          nullable: true
      additionalProperties: false
    EdFi.Ods.Api.Client.Models.StudentInterventionAssociationCohortReference:
      type: object
      properties:
        educationOrganizationId:
          type: integer
          format: int32
        identifier:
          type: string
          nullable: true
        link:
          $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.StudentInterventionAssociationCohortReferenceLink'
      additionalProperties: false
    EdFi.Ods.Api.Client.Models.StudentInterventionAssociationInterventionEffectivenessesItem:
      type: object
      properties:
        diagnosisDescriptor:
          type: string
          nullable: true
        gradeLevelDescriptor:
          type: string
          nullable: true
        improvementIndex:
          type: integer
          format: int32
          nullable: true
        interventionEffectivenessRatingType:
          type: string
          nullable: true
        populationServedType:
          type: string
          nullable: true
      additionalProperties: false
    EdFi.Ods.Api.Client.Models.StudentInterventionAssociationInterventionReferenceLink:
      type: object
      properties:
        href:
          type: string
          nullable: true
        rel:
          type: string
          nullable: true
      additionalProperties: false
    EdFi.Ods.Api.Client.Models.StudentInterventionAssociationInterventionReference:
      type: object
      properties:
        educationOrganizationId:
          type: integer
          format: int32
        identificationCode:
          type: string
          nullable: true
        link:
          $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.StudentInterventionAssociationInterventionReferenceLink'
      additionalProperties: false
    EdFi.Ods.Api.Client.Models.StudentInterventionAssociationStudentReferenceLink:
      type: object
      properties:
        href:
          type: string
          nullable: true
        rel:
          type: string
          nullable: true
      additionalProperties: false
    EdFi.Ods.Api.Client.Models.StudentInterventionAssociationStudentReference:
      type: object
      properties:
        link:
          $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.StudentInterventionAssociationStudentReferenceLink'
        studentUniqueId:
          type: string
          nullable: true
      additionalProperties: false
    EdFi.Ods.Api.Client.Models.StudentInterventionAssociation:
      type: object
      properties:
        _etag:
          type: string
          nullable: true
        cohortReference:
          $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.StudentInterventionAssociationCohortReference'
        diagnosticStatement:
          type: string
          nullable: true
        id:
          type: string
          nullable: true
        interventionEffectivenesses:
          type: array
          items:
            $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.StudentInterventionAssociationInterventionEffectivenessesItem'
          nullable: true
        interventionReference:
          $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.StudentInterventionAssociationInterventionReference'
        studentReference:
          $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.StudentInterventionAssociationStudentReference'
      additionalProperties: false
    EdFi.RtI.Core.DTOs.Composed.StudentInterventionsDTO+StudentInterventionAssociationDTO:
      type: object
      properties:
        associationModel:
          $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.StudentInterventionAssociation'
        intervention:
          $ref: '#/components/schemas/EdFi.Ods.Api.Client.Models.Intervention'
      additionalProperties: false
    EdFi.RtI.Core.DTOs.Composed.StudentInterventionsDTO:
      type: object
      properties:
        student:
          $ref: '#/components/schemas/EdFi.RtI.Core.DTOs.Student.StudentDTO'
        associations:
          type: array
          items:
            $ref: '#/components/schemas/EdFi.RtI.Core.DTOs.Composed.StudentInterventionsDTO+StudentInterventionAssociationDTO'
          nullable: true
      additionalProperties: false
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: http://localhost:5305/connect/authorize
          tokenUrl: http://localhost:5305/connect/token
          scopes:
            https://api.edgraph.com/auth/management: Management API